backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
google-generativeai==0.2.2
torch==2.0.0
pickle5==0.0.11
requests==2.31.0
